tracing module renamed to otel
