+++ internal/otel/trace_provider.go
@@ -1,4 +1,4 @@
-package tracing
+package otel
 
 import (
 	"context"
